removed NamedServiceEntity
