+++ geoide-domain/src/main/java/nl/idgis/geoide/commons/domain/FeatureType.java
@@ -6,9 +6,14 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-public final class FeatureType extends NamedServiceEntity {
+import nl.idgis.geoide.util.Assert;
+
+public final class FeatureType extends Entity {
 
 	private static final long serialVersionUID = 6708815788023537853L;
+	
+	private final Service service;
+	private final QName name;
 
 	@JsonCreator
 	public FeatureType (
@@ -16,7 +21,13 @@ public final class FeatureType extends NamedServiceEntity {
 			final @JsonProperty ("service") Service service, 
 			final @JsonProperty ("name") QName name,
 			final @JsonProperty ("label") String label) {
-		super (id, service, name, label);
+		super (id, label);
+		
+		Assert.notNull (service, "service");
+		Assert.notNull (name, "name");
+		
+		this.service = service;
+		this.name = name;	
 	}
 	
 	@JsonValue
@@ -24,10 +35,19 @@ public final class FeatureType extends NamedServiceEntity {
 		final ObjectNode n = JsonFactory.mapper ().createObjectNode ();
 		
 		n.put ("id", getId ());
-		n.put ("service", getService ().getId ());
-		n.put ("name", JsonFactory.mapper ().valueToTree (getName ()));
+		n.put ("service", service.getId ());
+		n.put ("name", JsonFactory.mapper ().valueToTree (name));
 		n.put ("label", getLabel ());
 		
 		return n;
+	}
+	
+	
+	public Service getService () {
+		return service;
+	}
+
+	public QName getName () {
+		return name;
 	}
 }
